pyproject.toml
[project]
name = "overthewire"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pwntools>=4.14.0",
    "python-dotenv>=1.0.1",
]
